+++ editor/views/record_sets_test.py
@@ -19,10 +19,10 @@ def test_find_joins():
             references=mlc.Source(uid="some_other_record_set/some_other_field"),
         ),
     ]
-    assert _find_joins(fields) == [
+    assert _find_joins(fields) == set([
         (("some_csv", "some_column"), ("some_record_set", "some_field")),
         (
             ("some_record_set", "some_field"),
             ("some_other_record_set", "some_other_field"),
         ),
-    ]
+    ])
